GetGame details endpoint was created
- Created IGDB and game service to extract logic
from endpoint controller
- Dtos created
- Result class created to organize returns from services

+++ GameStoreAPI/Models/GameInventory.cs
@@ -5,7 +5,7 @@ namespace GameStoreAPI.Models
     public class GameInventory
     {
         [Key]
-        public long IgdbId { get; set; } //KEY FROM IGDB
+        public int IgdbId { get; set; } //KEY FROM IGDB
         public decimal Price { get; set; }
         public decimal DiscountPrice { get; set; } = 0; 
         public int TotalSells { get; set; } = 0;
